Move assert_exit_status to process
